chore: rename ScheduleRowDeleteButton to ScheduleDeleteRowButton
