+++ web/src/components/RecommendedSchedule/RecommendedSchedule.tsx
@@ -1,7 +1,7 @@
 import Box from '@mui/joy/Box';
 import Table from '@mui/joy/Table';
 import Sheet from '@mui/joy/Sheet';
-import ScheduleDeleteRowButton from './ScheduleRowDeleteButton';
+import ScheduleDeleteRowButton from './ScheduleDeleteRowButton';
 import ScheduleCreateButton from './ScheduleCreateButton';
 import ScheduleDeleteButton from './ScheduleDeleteButton';
 import ScheduleEditButton from './ScheduleEditRowButton';
